requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
DAWG-Python==0.7.2
Deprecated==1.2.13
Django==3.2.12
docopt==0.6.2
flower==1.0.0
humanize==4.0.0
jarowinkler==1.0.2
joblib==1.1.0
kombu==5.2.4
Levenshtein==0.18.1
nltk==3.7
numpy==1.22.4
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.8
pytesseract==0.3.9
python-gettext==4.1
pytz==2022.1
rapidfuzz==2.0.11
redis==4.2.2
regex==2022.4.24
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tqdm==4.64.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
